版本更新

+++ contract/cache_handler_interface.go
@@ -6,8 +6,8 @@ type CacheHandlerInterface interface {
 	Get(key string) (any, error)
 	Set(key string, val any, expire time.Duration) error
 	Delete(key string) error
-	Increase(key string, step int) error
-	Decrease(key string, step int) error
+	Increase(key string, step int64) error
+	Decrease(key string, step int64) error
 	Expire(key string, dur time.Duration) error
 	Handler() any
 }
